+++ command/testing.go
@@ -4,6 +4,10 @@ import (
 	"errors"
 	"fmt"
 	"os/exec"
+	"reflect"
+
+	"github.com/AlecAivazis/survey/v2"
+	"github.com/AlecAivazis/survey/v2/core"
 
 	"github.com/cli/cli/api"
 	"github.com/cli/cli/context"
@@ -21,7 +25,7 @@ type CmdStubber struct {
 	Calls []*exec.Cmd
 }
 
-func InitCmdStubber() (*CmdStubber, func()) {
+func initCmdStubber() (*CmdStubber, func()) {
 	cs := CmdStubber{}
 	teardown := utils.SetPrepareCmd(createStubbedPrepareCmd(&cs))
 	return &cs, teardown
@@ -42,6 +46,57 @@ func createStubbedPrepareCmd(cs *CmdStubber) func(*exec.Cmd) utils.Runnable {
 		}
 		return cs.Stubs[call]
 	}
+}
+
+type askStubber struct {
+	Asks  [][]*survey.Question
+	Count int
+	Stubs [][]*QuestionStub
+}
+
+func initAskStubber() (*askStubber, func()) {
+	origSurveyAsk := SurveyAsk
+	as := askStubber{}
+	SurveyAsk = func(qs []*survey.Question, response interface{}, opts ...survey.AskOpt) error {
+		as.Asks = append(as.Asks, qs)
+		count := as.Count
+		as.Count += 1
+		if count >= len(as.Stubs) {
+			panic(fmt.Sprintf("more asks than stubs. most recent call: %v", qs))
+		}
+
+		// actually set response
+		stubbedQuestions := as.Stubs[count]
+		for i, sq := range stubbedQuestions {
+			q := qs[i]
+			if q.Name != sq.Name {
+				panic(fmt.Sprintf("stubbed question mismatch: %s != %s", q.Name, sq.Name))
+			}
+			if sq.Default {
+				defaultValue := reflect.ValueOf(q.Prompt).Elem().FieldByName("Default")
+				core.WriteAnswer(response, q.Name, defaultValue)
+			} else {
+				core.WriteAnswer(response, q.Name, sq.Value)
+			}
+		}
+
+		return nil
+	}
+	teardown := func() {
+		SurveyAsk = origSurveyAsk
+	}
+	return &as, teardown
+}
+
+type QuestionStub struct {
+	Name    string
+	Value   interface{}
+	Default bool
+}
+
+func (as *askStubber) Stub(stubbedQuestions []*QuestionStub) {
+	// A call to .Ask takes a list of questions; a stub is then a list of questions in the same order.
+	as.Stubs = append(as.Stubs, stubbedQuestions)
 }
 
 func initBlankContext(repo, branch string) {
